>> Resolution to issue 2: We should perform a roll call before the race starts. This puts the responsibility for identification on the boat captain. A fair call because many of the boats have names that can only be seen when they are within 5 feet of the committee boat. >> Resolution to issue 3: Adding the ability to add a new boat to the boat manifest during the boat selection phase. Simple! >> Resolution to issue 4: Adding the ability to change the boat assigned to a finish time via a drop down list on the Result Editor page >> Resolution to issue 5: Create a button called "Anonymous Finish" to the Finish Line screen Creates an blank finish line row with the time the button was pushed. When the user has time, they select the correct boat from a drop down list (user will have to remember the physical order of finish). Moving "send email" and "exit" buttons to the menu bar

+++ app/src/main/java/aad/finalproject/db/Result.java
@@ -16,7 +16,7 @@ public class Result {
     String resultsBoatFinishTime;
     String resultsDuration;
     String resultsAdjDuration;
-    double resultsPenalty;
+    int resultsPenalty;
     String resultsNote;
     int resultsPlace;
     double resultsPoints;
@@ -158,11 +158,11 @@ public class Result {
         this.resultsAdjDuration = resultsAdjDuration;
     }
 
-    public double getResultsPenalty() {
+    public int getResultsPenalty() {
         return resultsPenalty;
     }
 
-    public void setResultsPenalty(double resultsPenalty) {
+    public void setResultsPenalty(int resultsPenalty) {
         this.resultsPenalty = resultsPenalty;
     }
